requirements.txt
defusedxml==0.7.1
django-allauth==0.50.0
django-crispy-forms==1.14.0
docopt==0.6.2
greenlet==3.0.3
Jinja2==3.1.4
pillow==10.3.0
protonvpn-cli==2.2.11
psycopg2==2.9.9
python3-openid==3.2.0
pythondialog==3.5.3
requests-oauthlib==2.0.0
typing_extensions==4.12.2



